requirements.txt
allure-pytest>=2.13.5
allure-python-commons>=2.13.5
Faker>=37.1.0
pytest>=8.3.5
selenium>=4.29.0
attrs>=25.1.0
certifi>=2024.12.14
cffi>=1.17.1
colorama>=0.4.6
h11>=0.14.0
idna>=3.10
iniconfig>=2.0.0
outcome>=1.3.0.post0
packaging>=24.2
pluggy>=1.5.0
pycparser>=2.22
PySocks>=1.7.1
sniffio>=1.3.1
sortedcontainers>=2.4.0
trio>=0.28.0
trio-websocket>=0.11.1
typing_extensions>=4.12.2
tzdata>=2025.2
urllib3>=2.3.0
websocket-client>=1.8.0
wsproto>=1.2.0